The Message Cowboy is able yo choose another beast to ride.
Introducing Apache Camel.

    _______\\__
   (_. _ ._  _/ 
    '-' \__. /
         /  / 
        /  /    .--.  .--.
       (  (    / '' \/ '' \   "
        \  \_.'            \   )
        ||               _  './
         |\   \     ___.'\  /
           '-./   .'    \ |/ 
              \| /       )|\
               |/       // \\ 
               |\    __//   \\__
              //\\  /__/  mrf\__|
          .--_/  \_--.
         /__/      \__\
         

Setting the property messagecowboy.transport=camel will enable Apache
Camel instead of Mule ESB CE for transport matters. Camel and Mule
endpoint URIs are similar in structure, but not compatible.

The main reason to introduce another transport to Message Cowboy is to
give it more power. Even though Camel and Mule are largely overlapping
products, there are a lot of details and feature that are unique to each
of them. Giving the users of Message Cowboy a possibility to choose
transport engine will give them a feature range of transports not found
in any integration product known to man. (The union of features between
Camel and Mule is, well, impressive).

Future improvements could include running the two transports in
parallell and provice some briding between them. I.e. Camel inbound and
Mule outbound (or whatever).